Drop old removed tracks
ByteTracker library keeps old tracks around even when we don't
care about them anymore. This would cause a memory leak and possible
performance issues as well. Other mature reference implementation seem to cap
the number of tracks that are kept around in the removed list and pop the oldest
ones.

Signed-off-by: Phillip Kuznetsov <philkuz@gimletlabs.ai>